pyproject.toml
[project]
name = "madbot"
version = "0.1.0"
description = "Telegram bot with AI tendencies"
requires-python = ">=3.12"
dependencies = [
    "aiogram>=3.17,<4",
    "ai-sdk-python",
    "pydantic-settings>=2.0",
    "python-dotenv>=1.0",
]

[project.scripts]
madbot = "bot.main:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["bot"]
